requirements.txt
Django
django-registration
django-bootstrap4
django-bootstrap-breadcrumbs
django-filter
psycopg2
